refactor: #35 api 호출 시 에러 처리 훅으로 분리
