+++ src/api/theater/theater.api.ts
@@ -1,0 +1,40 @@
+import api from '@/api/api';
+import type { ApiResponse } from '@/types/api-response';
+import type { CinemaFormat } from '@/types/onboarding';
+import type { Theater } from '@/types/theater';
+
+// 영화관 목록 조회
+export interface GetTheatersParams {
+  type: CinemaFormat;
+  page: number;
+  size: number;
+}
+
+const getTheaters = async ({
+  type,
+  page = 1,
+  size = 10,
+}: GetTheatersParams): Promise<Theater[]> => {
+  const res = await api.get<ApiResponse<{ content: Theater[] }>>('/theaters', {
+    params: {
+      auditoriumType: type,
+      page,
+      size,
+    },
+  });
+
+  const content = res.data.data?.content;
+
+  if (!content) {
+    throw new Error('영화관 목록에 응답이 없습니다.');
+  }
+
+  return content;
+};
+
+// 좌석 배치도 조회
+
+// 평점/개수 포함된 좌석 배치도 조회
+
+// 상영관 상세 조회
+export { getTheaters };
